gogs: fix permission issues on startup
previously the Containerfile declared that the container should be
started as the internal 'gogs' user. This user does not have the
permission to write the SSH public key or start the SSH system service.

Therefore, the container is now started as root (internal) but then
drops privileges when executing the actual gogs web server.